requirements.txt
discord.py==0.10.0a0